# openapi-spec.json
{"openapi":"3.0.0","paths":{"/auth/login/google":{"post":{"operationId":"AuthController_loginGoogle","summary":"Login with Google (Do not forget to first authorize with your google id token.)","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginResponseDto"}}}}},"tags":["Authentication"],"security":[{"google-id-token":[]}]}},"/users/all":{"get":{"operationId":"UsersController_getAllUsers","summary":"Get all users.","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUsersResponseDto"}}}}},"tags":["Users"],"security":[{"jwt-access-token":[]}]}},"/users/me":{"get":{"operationId":"UsersController_getMyProfile","summary":"Get my profile","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUserResponseDto"}}}}},"tags":["Users"],"security":[{"jwt-access-token":[]}]}},"/users/update":{"post":{"operationId":"UsersController_updateUser","summary":"Updates a user.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserRequestDto"}}}},"responses":{"201":{"description":""}},"tags":["Users"],"security":[{"jwt-access-token":[]}]}},"/stratego/new-game":{"post":{"operationId":"StrategoController_newGame","summary":"Get all users.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewGameRequest"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewGameResponse"}}}}},"tags":["Stratego"],"security":[{"jwt-access-token":[]}]}},"/stratego/my-games":{"get":{"operationId":"StrategoController_myGames","summary":"Get my games.","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetMyGamesResponse"}}}}},"tags":["Stratego"],"security":[{"jwt-access-token":[]}]}},"/stratego/open-games":{"get":{"operationId":"StrategoController_openGames","summary":"Get open games.","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetOpenGamesResponse"}}}}},"tags":["Stratego"],"security":[{"jwt-access-token":[]}]}},"/stratego/game/{id}":{"delete":{"operationId":"StrategoController_deleteGame","summary":"Delete a game by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Stratego"],"security":[{"jwt-access-token":[]}]}}},"info":{"title":"Silly Things Games","description":"API Documentation for Silly Things Games Backend","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"jwt-access-token":{"scheme":"bearer","bearerFormat":"JWT","type":"http"},"google-id-token":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"UserDto":{"type":"object","properties":{"uid":{"type":"string"},"gid":{"type":"string"},"createdAt":{"type":"number"},"firstName":{"type":"string"},"lastName":{"type":"string"},"bio":{"type":"string"},"imageUrl":{"type":"string"}},"required":["uid","gid","createdAt","firstName","lastName","bio","imageUrl"]},"UserLoginResponseDto":{"type":"object","properties":{"accessToken":{"type":"string"},"user":{"$ref":"#/components/schemas/UserDto"},"isNewUser":{"type":"boolean"}},"required":["accessToken","user","isNewUser"]},"GetUsersResponseDto":{"type":"object","properties":{"users":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}}},"required":["users"]},"GetUserResponseDto":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/UserDto"}},"required":["user"]},"UpdateUserRequestDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"bio":{"type":"string"}},"required":["firstName","lastName","bio"]},"NewGameRequest":{"type":"object","properties":{"variation":{"type":"string","enum":["DRAGON","CLASSIC"]}},"required":["variation"]},"PieceDto":{"type":"object","properties":{"team":{"type":"number"},"originalTeam":{"type":"number"},"number":{"type":"number"},"x":{"type":"number"},"y":{"type":"number"}},"required":["team","originalTeam","number","x","y"]},"AffectedPieceDto":{"type":"object","properties":{"before":{"$ref":"#/components/schemas/PieceDto"},"after":{"$ref":"#/components/schemas/PieceDto"},"revealed":{"type":"boolean"},"killed":{"type":"boolean"}},"required":["before","after","revealed","killed"]},"MoveDto":{"type":"object","properties":{"piece":{"$ref":"#/components/schemas/AffectedPieceDto"},"piecesAffected":{"type":"array","items":{"$ref":"#/components/schemas/AffectedPieceDto"}}},"required":["piece","piecesAffected"]},"StrategoGameDto":{"type":"object","properties":{"variation":{"type":"string","enum":["DRAGON","CLASSIC"]},"gameId":{"type":"string"},"pieces":{"type":"array","items":{"$ref":"#/components/schemas/PieceDto"}},"previousMove":{"$ref":"#/components/schemas/MoveDto"},"currentTurn":{"type":"number"},"playerUids":{"type":"array","items":{"type":"string"}},"state":{"type":"string","enum":["WAITING","SETUP","PLAYING","FINISHED"]},"numTurns":{"type":"number"},"createdAt":{"type":"number"},"updatedAt":{"type":"number"}},"required":["variation","gameId","pieces","previousMove","currentTurn","playerUids","state","numTurns","createdAt","updatedAt"]},"NewGameResponse":{"type":"object","properties":{"game":{"$ref":"#/components/schemas/StrategoGameDto"}},"required":["game"]},"GetMyGamesResponse":{"type":"object","properties":{"games":{"type":"array","items":{"$ref":"#/components/schemas/StrategoGameDto"}}},"required":["games"]},"GetOpenGamesResponse":{"type":"object","properties":{"games":{"type":"array","items":{"$ref":"#/components/schemas/StrategoGameDto"}}},"required":["games"]}}}}